Import KDChart 1.1.4.
I had to disable support for QSA, since it didn't compile.  If you
want to reenable it, then remove the # in Makefile.am and remove #if 0
in KDChart.cpp.

In KDChartVectorTable.cpp, I had to disable the destructor since it
was defined in both the .h file and the .cpp file.  Apparently, the
author works with gcc version 3.x. :-)


svn path=/branches/koffice/1.6/koffice/; revision=600194

+++ kchart/kdchart/KDXMLTools.cpp
@@ -525,7 +525,7 @@ namespace KDXML {
                     ba[ i ] = r;
                 }
 
-                if( tempLength < (int)tempData.length() * 5 )
+                if( tempLength < tempData.length() * 5 )
                     tempLength = tempData.length() * 5;
                 QByteArray baunzip( tempLength );
                 ::uncompress( (uchar*) baunzip.data(), &tempLength,
